\docType{package}
\name{slidifyLibraries}
\alias{slidifyLibraries}
\alias{slidifyLibraries-package}
\title{slidifyLibraries contains external dependencies required by slidify}
\description{
  slidifyLibraries contains external dependencies required
  by slidify
}
\author{
  Ramnath Vaidyanathan <\url{http://www.slidify.org}>
}

